+++ Package.swift
@@ -1,20 +1,28 @@
-// swift-tools-version:5.0
+// swift-tools-version:5.9
 import PackageDescription
 
 let package = Package(
     name: "FLAnimatedImage",
     platforms: [
-        .iOS(.v9)
+        .iOS(.v14)
     ],
     products: [
-        .library(name: "FLAnimatedImage", targets: ["FLAnimatedImage"]),
+        .library(
+            name: "FLAnimatedImage",
+            targets: ["FLAnimatedImage"]
+        )
+    ],
+    dependencies: [
+        .package(url: "https://github.com/SDWebImage/libwebp-Xcode", from: "1.3.2")
     ],
     targets: [
         .target(
             name: "FLAnimatedImage",
+            dependencies: [
+                .product(name: "libwebp", package: "libwebp-xcode")
+            ],
             path: "FLAnimatedImage",
             exclude: [ "Info.plist" ],
-            sources: [ "FLAnimatedImageView.m", "FLAnimatedImage.m" ],
             publicHeadersPath: "include",
             cSettings: [
                 .headerSearchPath("include")
