--- Podfile
@@ -1,20 +1,0 @@
-workspace 'FLAnimatedImage.xcworkspace'
-
-# Uncomment the next line to define a global platform for your project
- platform :ios, '14.0'
-
-target 'FLAnimatedImage' do
-  project 'FLAnimatedImage.xcodeproj'
-  # Comment the next line if you don't want to use dynamic frameworks
-  use_frameworks!
-
-  # Pods for FLAnimatedImage
-  pod 'libwebp'
-end
-
-target 'FLAnimatedImageDemo' do
-  project 'FLAnimatedImageDemo.xcodeproj'
-  use_frameworks!
-
-  pod 'libwebp'
-end
